+++ wait_for_postgres.sh
@@ -3,6 +3,7 @@
 set -e
 
 export PGHOST=$HOMS_DB_HOST
+export PGPORT=$HOMS_DB_PORT
 export PGUSER=$HOMS_DB_USER
 export PGPASSWORD=$HOMS_DB_PASSWORD