+++ CHANGELOG.md
@@ -22,6 +22,7 @@ v1.6.0 [unreleased]
 -   [#232](https://github.com/latera/homs/pull/232) [DEPRECATION] Deprecate Activiti usage.
 -   [#233](https://github.com/latera/homs/pull/233) Run DB on port specified in `.env` file.
 -   [#231](https://github.com/latera/homs/pull/231) Add `multi` option to select_table.
+-   [#240](https://github.com/latera/homs/pull/240) Update ruby up to 2.6.0.
 
 ### Bugfixes
 -   [#194](https://github.com/latera/homs/pull/194) Set limit on the number of choices in lookup select from sql requests.