+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+v1.6.1 [unreleased]
+-------------------
+
+### Bugfixes
+-   [#241](https://github.com/latera/homs/pull/241) Fix chevron reaction on task collapse.
+
 v1.6.0 [2019-01-29]
 -------------------